server/api/requirements.txt
asn1crypto==0.24.0
attrs==19.3.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2018.1.18
cffi==1.14.0
chardet==3.0.4
click==7.1.1
command-not-found==0.3
config==0.5.0.post0
console-log==0.2.10
cryptography==2.8
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
elasticsearch==7.5.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-SQLAlchemy==2.4.1
gevent==1.5.0
gevent-websocket==0.10.1
greenlet==0.4.15
httplib2==0.9.2
idna==2.6
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
marshmallow==3.4.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
paramiko==2.7.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psutil==5.7.0
psycopg2-binary==2.8.4
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
pyrsistent==0.16.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.6.1
python-debian==0.1.32
python-terraform==0.10.1
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.7.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
SQLAlchemy==1.3.13
system-service==0.3
systemd-python==234
texttable==1.6.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
websocket-client==0.57.0
Werkzeug==1.0.1
wsgigzip==0.1.4
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
